Move item popup and armory out of the initial chunk
