+++ src/app/armory/LazyArmory.ts
@@ -1,0 +1,3 @@
+import { lazy } from 'react';
+
+export default lazy(() => import('./Armory' /* webpackChunkName: "armory" */));
